ci: move test config up one level
Some of our tests are placed with their submodules. This means that settings which should affect all tests have to be placed one level higher.
